Tried to make 2D ArrayList of trees and changed to different folder
Tried to get an array list of an array list of trees to add trees in a nested for loop...still haven' figured it out, but I'm taking a break now so time to save! There's also some new comments and a few touches-up, like I added a new tab for misc stuff. Also stupidly edited the old, not-in-github folder so I switched them around..hopefully this works ok. The one with A is the correct one.