More stuff in coin class and backwards loops
Added coin scroll, finished coin collide, added disappear and reappear plus coin x constraint in constructor, a way to score that's really weird because every time you score it's twice as much as it actually is (so I made it move up by 0.5 and typecasted to an int). Added global variable for scroll speed, ran for loops backwards, and some more comments/psuedo code. Next order of businesss is levels and controlled randomness!