Did a lot of stuff but got rid of most of it
Why doesn't the collide work with coins?? It did before...
changed cases so now there's a screen between the title page and the game start that directs you to places. Tried to get a hit animation but gave up because I need a timer I think for what I'm trying o do  and though I added one its not working. added a reset button for player position and stats.