Fixed a lot of stuff-reset button!
Fixed the issues with scrolling and acceleration for the screen and the player, and made a reset button. It only works when you lsoe though; still need reset for winning. Also, when it resets the old trees and coins don't disapear....I'm gonna try moving the for loops in the setup t the gamerunning area